"all the functionalityes modified and completed with cart controller also"

+++ controller/productController.js
@@ -2,14 +2,30 @@ import { ProductModel } from "../model/product.model.js";
 import { catchAsyncError } from "../utils/catchAsyncErrors.js";
 import { AppError } from "../utils/ApiError.js";
 import slugify from "slugify";
-
+import fs from 'fs/promises';
+ 
 export const addProduct = catchAsyncError(async (req, res, next) => {
-    const images = req.files.prodImages.map((ele) => ele.filename);
-    req.body.images = images;
-    req.body.slug = slugify(req.body.title);
-    const addProduct = new ProductModel(req.body);
-    await addProduct.save();
-    res.status(201).json({ message: "success", addProduct });
+    if (!req.files || !req.files.length) {
+        return next(new AppError('Please upload at least one image', 400));
+    }
+    try {
+        const uploadPromises = req.files.map(file => uploadToCloudinary(file));
+        const uploadedImages = await Promise.all(uploadPromises);
+        req.body.images = uploadedImages;
+        req.body.slug = slugify(req.body.title);
+
+        const addProduct = new ProductModel(req.body);
+        await addProduct.save();
+
+        // Cleanup temporary files
+        await Promise.all(req.files.map(file => fs.unlink(file.path)));
+
+        res.status(201).json({ message: "success", addProduct });
+    } catch (error) {
+        // Cleanup temporary files in case of error
+        await Promise.all(req.files.map(file => fs.unlink(file.path).catch(() => {})));
+        return next(new AppError('Error uploading images: ' + error.message, 500));
+    }
 });
 
 export const getAllProduct = catchAsyncError(async (req, res, next) => {
@@ -25,21 +41,49 @@ export const getProductById = catchAsyncError(async (req, res, next) => {
 
 export const updateProduct = catchAsyncError(async (req, res, next) => {
     const productId = req.params.id;
+    const product = await ProductModel.findById(productId);
+
+    if (!product) {
+        return next(new AppError("Product was not found", 404));
+    }
+
     if (req.body.title) {
         req.body.slug = slugify(req.body.title);
+    } 
+
+    if (req.files && req.files.length > 0) {
+        try {
+            // Upload new images to Cloudinary
+            const uploadPromises = req.files.map(file => uploadToCloudinary(file));
+            const uploadedImages = await Promise.all(uploadPromises);
+            req.body.images = uploadedImages;
+
+            // Cleanup temporary files
+            await Promise.all(req.files.map(file => fs.unlink(file.path)));
+        } catch (error) {
+            // Cleanup temporary files in case of error
+            await Promise.all(req.files.map(file => fs.unlink(file.path).catch(() => {})));
+            return next(new AppError('Error uploading images: ' + error.message, 500));
+        }
     }
+
     const updatedProduct = await ProductModel.findByIdAndUpdate(productId, req.body, {
         new: true
     });
-    if (updatedProduct) {
-        res.status(200).json({ message: "success", updatedProduct });
-    } else {
-        next(new AppError("Product was not found", 404));
-    }
+
+    res.status(200).json({ message: "success", updatedProduct });
 });
 
 export const deleteProduct = catchAsyncError(async (req, res, next) => {
     const productId = req.params.id;
+    const product = await ProductModel.findById(productId);
+
+    if (!product) {
+        return next(new AppError("Product not found", 404));
+    }
+
+    // No need to delete images from Cloudinary as they are managed there
+
     await ProductModel.findByIdAndDelete(productId);
     res.status(204).json({ message: "success" });
 });
